Re-organize profile fixtures a bit
