+++ src/test/fixtures/profiles/gecko-profile.js
