+++ src/test/store/actions.test.js
@@ -1,12 +1,13 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+// @flow
 
 import { storeWithProfile } from '../fixtures/stores';
 import * as ProfileViewSelectors from '../../reducers/profile-view';
 import * as TimelineSelectors from '../../reducers/timeline-view';
 import * as UrlStateSelectors from '../../reducers/url-state';
-import { getProfileWithMarkers } from './fixtures/profiles';
+import { getProfileWithMarkers } from '../fixtures/profiles/make-profile';
 
 import {
   changeCallTreeSearchString,
