+++ src/test/store/receive-profile.test.js
@@ -15,7 +15,7 @@ import {
 } from '../../actions/receive-profile';
 
 import preprocessedProfile from '../fixtures/profiles/profile-2d-canvas.json';
-import exampleProfile from '../fixtures/profiles/example-profile';
+import getGeckoProfile from '../fixtures/profiles/gecko-profile';
 
 describe('actions/receive-profile', function() {
   /**
@@ -61,7 +61,7 @@ describe('actions/receive-profile', function() {
       clock = sinon.useFakeTimers();
 
       geckoProfiler = {
-        getProfile: () => Promise.resolve(exampleProfile()),
+        getProfile: () => Promise.resolve(getGeckoProfile()),
         getSymbolTable: () => Promise.resolve(),
       };
       window.geckoProfilerPromise = Promise.resolve(geckoProfiler);
@@ -129,7 +129,7 @@ describe('actions/receive-profile', function() {
     const fetch200Response = {
       ok: true,
       status: 200,
-      json: () => Promise.resolve(exampleProfile()),
+      json: () => Promise.resolve(getGeckoProfile()),
     };
 
     beforeEach(function() {
@@ -243,7 +243,7 @@ describe('actions/receive-profile', function() {
     const fetch200Response = {
       ok: true,
       status: 200,
-      json: () => Promise.resolve(exampleProfile()),
+      json: () => Promise.resolve(getGeckoProfile()),
     };
 
     beforeEach(function() {
